Fix video flash issue at the end of playing video.
BZ:141685

MDS calls HWC to disable overlay when enter into unpreparing state.
This will make video window to be black, which is not expected.
For protected video, operations on overlay plane will not change,
For clear video, overlay will not be disabled in unpreparing state.

+++ common/devices/ExternalDevice.cpp
@@ -266,7 +266,7 @@ void ExternalDevice::setRefreshRate(int hz)
     if (hz == 0 && (mode.type & DRM_MODE_TYPE_PREFERRED))
         return;
 
-    if (hz == mode.vrefresh)
+    if (hz == (int)mode.vrefresh)
         return;
 
     ITRACE("changing refresh rate from %d to %d", mode.vrefresh, hz);
